Update EPApprovalMapMaint.cs for reason fields
On the screen for approval maps, the base event for EPRule_RowSelected will leave the reason fields hidden for custom screens that might leverage approval comments.  The previous version of this file reset the fields' visibility from scratch.  This change will leverage the base event's setting of visibility to hide the fields and then toggle back to visible for screens added here.  When this updated approach is used on multiple customization projects, they will not reset visibility leaving only the last project loaded working but rather allow all custom screens to retain visibility of the fields.