Update AUGOrderEntry_ApprovalWorkflow.cs
In the Prefetch, requestApproval was set to (bool)setup.GetBoolean(0) which is null until the Request Approval field is checked (and subsequently unchecked if desired).  This "unset" value returns as null which cannot be cast to bool.  This breaks the entire workflow for the screen because this graph extension layers on top of the main workflow graph extension.  The fix is simply to cast the field properly and return if that value equals true.  i.e. ((bool?)setup.GetBoolean(0) == true);